# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Purple Squad API
  version: v1
  description: API for the Purple Squad home services platform.
paths:
...
  /api/v1/admin/bookings/{booking_id}/cancel/:
    post:
      operationId: api_v1_admin_bookings_cancel_create
      summary: Run admin booking operation
      parameters:
      - in: path
        name: booking_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingOperationRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BookingOperationRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BookingOperationRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
          description: ''
...
  /api/v1/bookings/{id}/cancel/:
    post:
      operationId: api_v1_bookings_cancel_create
      description: Cancels an authenticated customer's booking when policy and status
        allow it.
      summary: Cancel customer booking
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingOperationRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BookingOperationRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BookingOperationRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
          description: ''
  /api/v1/bookings/{id}/reschedule/:
    post:
      operationId: api_v1_bookings_reschedule_create
      description: Moves an authenticated customer's booking to another available
        slot when policy allows it.
      summary: Reschedule customer booking
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRescheduleRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BookingRescheduleRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BookingRescheduleRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
          description: ''
...
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        booking_number:
          type: string
          maxLength: 32
        service:
          type: object
          additionalProperties: {}
          readOnly: true
        address_snapshot: {}
        service_date:
          type: string
          format: date
        time_slot:
          type: object
          additionalProperties: {}
          readOnly: true
        problem_description:
          type: string
        subtotal:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        discount_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        tax_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        total_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        advance_required:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        advance_paid:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        balance_due:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        balance_collected:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        payment_status:
          $ref: '#/components/schemas/PaymentStatusEnum'
        booking_status:
          $ref: '#/components/schemas/BookingStatusEnum'
        customer_notes:
          type: string
        admin_notes:
          type: string
        confirmed_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        cancelled_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        status_history:
          type: array
          items:
            $ref: '#/components/schemas/BookingStatusHistory'
          readOnly: true
      required:
      - address_snapshot
      - advance_required
      - balance_due
      - booking_number
      - created_at
      - id
      - problem_description
      - service
      - service_date
      - status_history
      - subtotal
      - time_slot
      - total_amount
      - updated_at
...
    BookingOperationRequest:
      type: object
      properties:
        notes:
          type: string
    BookingRescheduleRequest:
      type: object
      properties:
        slot_id:
          type: string
          format: uuid
        notes:
          type: string
      required:
      - slot_id
    BookingStatusEnum:
      enum:
      - PENDING_PAYMENT
      - PAYMENT_FAILED
      - CONFIRMED
      - TECHNICIAN_ASSIGNED
      - TECHNICIAN_EN_ROUTE
      - IN_PROGRESS
      - COMPLETED
      - CANCELLED
      - REFUND_PENDING
      - REFUNDED
      type: string
      description: |-
        * `PENDING_PAYMENT` - Pending payment
        * `PAYMENT_FAILED` - Payment failed
        * `CONFIRMED` - Confirmed
        * `TECHNICIAN_ASSIGNED` - Technician assigned
        * `TECHNICIAN_EN_ROUTE` - Technician en route
        * `IN_PROGRESS` - In progress
        * `COMPLETED` - Completed
        * `CANCELLED` - Cancelled
        * `REFUND_PENDING` - Refund pending
        * `REFUNDED` - Refunded
    BookingStatusHistory:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        from_status:
          type: string
          maxLength: 32
        to_status:
          type: string
          maxLength: 32
        notes:
          type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - to_status
...
    PaymentStatusEnum:
      enum:
      - UNPAID
      - PARTIALLY_PAID
      - PAID
      - FAILED
      - REFUNDED
      type: string
      description: |-
        * `UNPAID` - Unpaid
        * `PARTIALLY_PAID` - Partially paid
        * `PAID` - Paid
        * `FAILED` - Failed
        * `REFUNDED` - Refunded
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT